SIM Application Toolkit: Porting UI to QML
  * Pass two: all widgets except text / char input
  * Added constructor to StkDialog to use a qml view instead of a widget
  * Display qml views from StkAgentService where available

+++ src/stkwidget.h
@@ -26,7 +26,7 @@ signals:
     void rejected();
     void goBack();
     void endSession();
-    void itemSelected(uchar selection);
+    void itemSelected(int selection);
     void textEntered(QString text);
 };
